fully qualify other files, pull client out of other files, register method to register self for client tree
